feat: Introduce configuration management and logging setup
Improve the application's configuration and logging mechanisms.

* Extracted configuration loading into a separate module (`config.py`) for better organization and reusability.
* Implemented a hierarchical configuration loading system with environment files and variables.
* Enhanced logging setup with customizable log levels, quiet mode, and noisy logger suppression.
* Updated main entry point to utilize the new configuration and logging features.
